requirements_rpi.txt
numpy
opencv-python
picamera
pyserial
tqdm
pyyaml
pycoral